Update RecipeApp.java
Managed to return a list of recipes by time and ingredient.

Added a condition to the remove task, to return the recipe if it exists and confirm if you wish to delete it. While the code confirms the deletion it's still on the txt file. Can't figure out why. 

There are still a lot of inconsistencies, with random prints poping up,  but at least it's something.